% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_dir.R
\name{clean_model_dir}
\alias{clean_model_dir}
\title{Clean ezcox Model File Directory}
\usage{
clean_model_dir(model_dir = file.path(tempdir(), "ezcox"))
}
\arguments{
\item{model_dir}{a path for storing model results.}
}
\value{
nothing
}
\description{
Clean ezcox Model File Directory
}
\examples{
\donttest{
clean_model_dir()
}
}
